d/inventrack/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pymysql
python-multipart